feat: added downsampling and pagination
