+++ webapp/metrics/helper.py
@@ -167,3 +167,86 @@ def transform_metrics(metrics, metrics_response, snaps):
             metrics["buckets"] = metric["buckets"]
 
     return metrics
+
+
+def lttb_select_indices(values, target_size):
+    """Selects indices using the LTTB algorithm for downsampling, treating None as 0."""
+    n = len(values)
+    if n <= target_size:
+        return list(range(n))
+
+    # Initialize bucket size
+    bucket_size = (n - 2) / (target_size - 2)
+    indices = []
+
+    current_bucket_start = 0
+    for i in range(1, target_size - 1):
+        next_bucket_start = min(math.ceil((i + 1) * bucket_size), n - 1)
+
+        max_area = 0
+        max_area_idx = current_bucket_start
+
+        point1 = (current_bucket_start, values[current_bucket_start] if values[current_bucket_start] is not None else 0)
+        point2 = (next_bucket_start, values[next_bucket_start] if values[next_bucket_start] is not None else 0)
+
+        # Calculate the area for each valid index between current and next bucket
+        for j in range(current_bucket_start + 1, next_bucket_start):
+            val_j = values[j] if values[j] is not None else 0
+
+            # Area of triangle formed by point1, point2, and the current point
+            area = abs(
+                (point1[0] - point2[0]) * (val_j - point1[1])
+                - (point1[0] - j) * (point2[1] - point1[1])
+            )
+            if area > max_area:
+                max_area = area
+                max_area_idx = j
+
+        indices.append(max_area_idx)
+        current_bucket_start = next_bucket_start
+
+    indices.append(n - 1)
+    return indices
+
+def normalize_series(series, bucket_count):
+    """Ensure all value arrays in the series have the same size by padding with 0s."""
+    for item in series:
+        values = item['values']
+        # If the series has no values, fill it with 0s
+        if not values:
+            item['values'] = [0] * bucket_count
+        # Extend the values with 0 if they are shorter than the bucket count
+        elif len(values) < bucket_count:
+            item['values'].extend([0] * (bucket_count - len(values)))
+
+
+def downsample_series(buckets, series, target_size):
+    """Downsample each series in the data, treating None as 0."""
+    downsampled_buckets = []
+    downsampled_series = []
+
+    # Handle case where series is empty
+    if not series:
+        return buckets[:target_size], []
+
+    bucket_count = len(buckets)
+    # Normalize series first to make sure all series have the same length
+    normalize_series(series, bucket_count)
+
+    # Downsample each series independently
+    for item in series:
+        name = item['name']
+        values = item['values']
+        
+        selected_indices = lttb_select_indices(values, target_size)
+        
+        # Collect the downsampled buckets and values based on the selected indices
+        downsampled_buckets = [buckets[i] for i in selected_indices]
+        downsampled_values = [values[i] if values[i] is not None else 0 for i in selected_indices]
+
+        downsampled_series.append({
+            'name': name,
+            'values': downsampled_values
+        })
+
+    return downsampled_buckets, downsampled_series
